Time: 51 ms (94.3%), Space: 11.1 MB (74.43%) - LeetHub

+++ 1781-sum-of-beauty-of-all-substrings/1781-sum-of-beauty-of-all-substrings.cpp
@@ -11,8 +11,8 @@ public:
                 int maxi = 0;
                 for(int f:freq){
                     if(f>0){
-                        mini = min(mini,(int)f);
-                        maxi = max(maxi,(int)f);
+                        mini = min(mini,f);
+                        maxi = max(maxi,f);
                     }
                 }
                 ans+=(maxi-mini);
